Add new queries, that are aligned with current objects.
restaurantSummariesQuery, is used for fetching RestaurantSummaryModel objects, this objects will be used in AllRestaurants tab,

restaurantDetailQuery, is used for fetching RestaurantModel objects, this objects will be used in restaurants detail view screens.

restaurantSummariesWithIdsQuery, is used for fetching RestaurantSummaryModel objects by restaurant id's. This will be used to fetch favorited restaurants on favorites tab.